+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 - Don't emit empty `interfaces` in type hints.
 - Don't `try/except` imports in `.pyi` files.
+- Don't use `typing.Optional` in generic type parameters.
 
 ### Fixed
 - Fixed support for sub-interpreters.